+++ src/main/java/com/mdh/devtable/reservation/infra/persistence/ShopReservationDateTimeRepository.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.reservation.infra;
+package com.mdh.devtable.reservation.infra.persistence;
 
 import com.mdh.devtable.reservation.ShopReservationDateTime;
 import org.springframework.data.jpa.repository.JpaRepository;
